Add database entities and repository for plan management, including TrainingDay, ExerciseEntry, and SetEntry

+++ src/main/kotlin/de/powerplan/plans/infrastructure/adapters/rest/requests/CreatePlanRequest.kt
@@ -65,7 +65,7 @@ data class CreateSetEntryRequest(
     val rpe: Double? = null,
     val minRpe: Double? = null,
     val maxRpe: Double? = null,
-    val percent1RM: Int? = null
+    val percent1RM: Double? = null
 ) {
     fun toCommand() = CreateSetEntryCommand(
         fixedReps = fixedReps,
